refactor all folders and files
