+++ src/Pages/LandingPage.tsx
@@ -2,7 +2,7 @@
 import React, { useState } from 'react';
 import './LandingPage.css';
 import { Link } from 'react-router-dom';
-import Navbar from '../components/Navbar';
+import Navbar from '../components/common/Navbar/Navbar';
 
 function LandingPage() {
   const [showNavbar] = useState(false);
