+++ src/Pages/LoginPage.tsx
@@ -6,7 +6,7 @@ import './LoginPage.css';
 
 import { Card, Input, Button, Typography } from '@material-tailwind/react';
 
-import { BASE_URL } from '../utility/endpoints';
+import { BASE_URL } from '../utils/endpoints';
 
 function LoginPage() {
   const [email, setEmail] = useState('');
