+++ src/components/common/Navbar/Navbar.css
@@ -3,7 +3,7 @@
   height: 300px;
   left: 0px;
   top: 0px;
-  background: url('../assets/Navbar-plane.png') center/cover no-repeat;
+  background: url('../../../assets/Navbar-plane.png') center/cover no-repeat;
   /* width: 100vw; */
 }
 .dark {
