+++ src/components/common/SearchBar/SearchBar.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Input, Select, Option, Button, Radio } from '@material-tailwind/react';
 
-import DatePicker from './DatePicker';
+import DatePicker from '../../utilities/DatePicker';
 
 // import { DataSourceContext } from '../context/DataSource.context'
