+++ src/components/flights/FlightCards.tsx
@@ -1,14 +1,14 @@
 import React, { useContext } from 'react';
 import { Rating, Button, Typography } from '@material-tailwind/react';
-import { CartContext } from '../context/Cart.context';
-import firstIcon from '../assets/airline-icon-first.png';
-import secondIcon from '../assets/airline-icon-second.png';
-import bagsIcon from '../assets/bags.png';
-import separator from '../assets/separator.png';
+import { CartContext } from '../../context/Cart.context';
+import firstIcon from '../../assets/airline-icon-first.png';
+import secondIcon from '../../assets/airline-icon-second.png';
+import bagsIcon from '../../assets/bags.png';
+import separator from '../../assets/separator.png';
 
 import './FlightCards.css';
 import TicketFlightInfo from './TicketFlightInfo';
-import { Flight } from '../types';
+import { Flight } from '../../types';
 
 interface FlightCardsProps {
   flight: Flight[];
