+++ src/context/Cart.context.tsx
@@ -2,7 +2,7 @@ import React, { createContext, useState, useEffect } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 
-import { PURCHASE_API, FLIGHTS_ALL_API } from '../utility/endpoints';
+import { PURCHASE_API, FLIGHTS_ALL_API } from '../utils/endpoints';
 
 const fetchingData = async () => {
   const response = await axios.get(FLIGHTS_ALL_API);
