+++ src/main.tsx
@@ -6,7 +6,6 @@ import './index.css';
 import { AuthProviderWrapper } from './context/Auth.context.js';
 import { CartContextProvider } from './context/Cart.context.js';
 import { DataSourceProvider } from './context/DataSource.context.js';
-import { ThemeProvider } from './context/ThemeContext.js';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
